Move SKR tests to KEB repository
