+++ tests/fast-integration/eventing-test/kcp/client.js
@@ -5,7 +5,7 @@ const {
   getEnvOrThrow,
   debug,
   wait,
-} = require('../utils');
+} = require('../../utils');
 const {inspect} = require('util');
 
 class KCPConfig {
