+++ tests/fast-integration/eventing-test/kyma-environment-broker/helpers.js
@@ -1,4 +1,4 @@
-const {wait, debug} = require('../utils');
+const {wait, debug} = require('../../utils');
 const {expect} = require('chai');
 const fs = require('fs');
 const os = require('os');
