--- tests/fast-integration/skr-aws-upgrade-integration/.env.template
@@ -1,10 +1,0 @@
-KEB_HOST=
-KEB_CLIENT_ID=
-KEB_CLIENT_SECRET=
-KEB_TOKEN_URL=
-KEB_GLOBALACCOUNT_ID=
-KEB_SUBACCOUNT_ID=
-KEB_USER_ID=
-KEB_PLAN_ID=
-GARDENER_KUBECONFIG=
-AL_SERVICE_KEY=
