--- tests/fast-integration/skr-nightly/.env.template
@@ -1,17 +1,0 @@
-KEB_HOST=
-KEB_CLIENT_ID=
-KEB_CLIENT_SECRET=
-KEB_TOKEN_URL=
-KEB_GLOBALACCOUNT_ID=
-KEB_SUBACCOUNT_ID=
-KEB_USER_ID=
-KEB_PLAN_ID=
-GARDENER_KUBECONFIG=
-AL_SERVICE_KEY=
-KCP_KEB_API_URL=
-KCP_OIDC_ISSUER_URL=
-KCP_GARDENER_NAMESPACE=
-KCP_TECH_USER_LOGIN=
-KCP_TECH_USER_PASSWORD=
-KCP_OIDC_CLIENT_ID=
-KCP_OIDC_CLIENT_SECRET=
