--- tests/fast-integration/skr-test/.env.template
@@ -1,35 +1,0 @@
-KEB_HOST=
-KEB_CLIENT_ID=
-KEB_CLIENT_SECRET=
-KEB_GLOBALACCOUNT_ID=
-KEB_SUBACCOUNT_ID=
-KEB_USER_ID=
-KEB_PLAN_ID=
-KEB_TOKEN_URL=
-GARDENER_KUBECONFIG=
-AL_SERVICE_KEY=
-
-KCP_KEB_API_URL=
-KCP_GARDENER_NAMESPACE=
-KCP_OIDC_ISSUER_URL=
-KCP_OIDC_CLIENT_ID=
-KCP_OIDC_CLIENT_SECRET=
-KCP_TECH_USER_LOGIN=
-KCP_TECH_USER_PASSWORD=
-KCP_MOTHERSHIP_API_URL=
-KCP_KUBECONFIG_API_URL=
-
-# required for non-provisioning mode
-SKIP_PROVISIONING=
-INSTANCE_ID=
-
-BTP_OPERATOR_CLIENTID=
-BTP_OPERATOR_CLIENTSECRET=
-BTP_OPERATOR_URL=
-BTP_OPERATOR_TOKENURL=
-
-# optional 
-KCP_OAUTH2_CLIENT_SECRET=
-KCP_OAUTH2_CLIENT_ID=
-KCP_OAUTH2_ISSUER_URL=
-GARDENER_SHOOT_TEMPLATE=
