--- tests/fast-integration/skr-test/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-  ...require('./helpers'),
-  ...require('./provision/provision-skr'),
-  ...require('./oidc/index'),
-  ...require('./machine-type/index'),
-  ...require('./btp-manager-secret/index'),
-};
